+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## NEXT
  - Removed useless lifetime parameter from `SessionConfig::with_database()`.
- - Changed return type of `ConnectionConfig::with_encryption_trust_any_certificate() ` to from `Result<Self, Error>` to `Self`.
+ - Changed return type of `ConnectionConfig::with_encryption_trust_any_certificate() ` from `Result<Self, Error>` to `Self`.
+ - Add support for Bolt 5.2, which adds notification filtering.
 
 ## 0.0.2
  - Update dependencies.